move code to own file
