+++ src/TypealizR.SourceGenerators/StringLocalizer/SourceGenerator.cs
@@ -1,71 +1,16 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Text;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Diagnostics;
 using TypealizR.SourceGenerators.Extensions;
 
 namespace TypealizR.SourceGenerators.StringLocalizer;
 
 [Generator]
-public class SourceGenerator : IIncrementalGenerator
+public partial class SourceGenerator : IIncrementalGenerator
 {
-    private class Settings
-    {
-        public Settings(string? projectDirectory, string? rootNamespace, IDictionary<string, DiagnosticSeverity> severityConfig)
-        {
-            RootNamespace = rootNamespace ?? "";
-			SeverityConfig = severityConfig;
-			ProjectDirectory = new DirectoryInfo(projectDirectory);
-        }
-
-        public DirectoryInfo ProjectDirectory { get; }
-        public string RootNamespace { get; }
-		public IDictionary<string, DiagnosticSeverity> SeverityConfig { get; }
-
-		public static Settings From(AnalyzerConfigOptions options)
-		{
-			if (!options.TryGetValue("build_property.msbuildprojectdirectory", out var projectDirectory))
-			{
-				options.TryGetValue("build_property.projectdir", out projectDirectory);
-			}
-
-			options.TryGetValue("build_property.rootnamespace", out var rootNamespace);
-
-			var severityConfig = ReadSeverityConfig(options);
-
-			return new(
-				projectDirectory: projectDirectory ?? Guid.NewGuid().ToString(),
-				rootNamespace: rootNamespace ?? Guid.NewGuid().ToString(),
-				severityConfig: severityConfig
-			);
-		}
-
-		private static IDictionary<string, DiagnosticSeverity> ReadSeverityConfig(AnalyzerConfigOptions options)
-		{
-			var severityConfig = new Dictionary<string, DiagnosticSeverity>();
-			foreach (var diagnosticId in new[] { "TR0001", "TR0002", "TR0003", "TR0004" }) //TODO: Do not violate open-closed-principle here
-			{
-				if (options.TryGetValue($"dotnet_diagnostic_{diagnosticId.ToLower()}_severity", out var rawValue))
-				{
-					if (Enum.TryParse<DiagnosticSeverity>(rawValue, true, out var severity))
-					{
-						severityConfig[diagnosticId] = severity;
-					}
-					else
-					{
-						//should we better error here?
-					}
-				}
-			}
-
-            return severityConfig;
-		}
-	}
 
     public void Initialize(IncrementalGeneratorInitializationContext context)
     {
